Update sortable-item.js
Changed index management for 'asSortableItem' directive. When the item list is paginated, the $index variable on the scope refers to the first page elements. The only way to refer to the right item is to add a 'ngModel' directive (with a bit of redundancy), without using the $index at all.